+++ test/web/mastodon_api/views/account_view_test.exs
@@ -94,10 +94,9 @@ defmodule Pleroma.Web.MastodonAPI.AccountViewTest do
     user = insert(:user)
 
     notification_settings = %{
-      followers: true,
-      follows: true,
-      non_followers: true,
-      non_follows: true,
+      from_followers: true,
+      from_following: true,
+      from_strangers: true,
       privacy_option: false
     }